feat: allow addition of webpack plugins. fixes #641
This is a fairly simple extension point that depends upon a statically
named file hlx.webpack.customize.js. This is expected to export
a function named extraPlugins that accepts the hedy config and the
webpack config and returns an array of plugins.

This is an alternative to #642 which provides a narrower extension point.